+++ src/frontend/src/features/rooms/livekit/components/controls/Options/OptionsButton.tsx
@@ -1,22 +1,38 @@
 import { useTranslation } from 'react-i18next'
 import { RiMore2Line } from '@remixicon/react'
-import { Button } from '@/primitives'
-import { OptionsMenu } from '@/features/rooms/livekit/components/controls/Options/OptionsMenu.tsx'
-import { MenuTrigger } from 'react-aria-components'
+import { Button, Menu } from '@/primitives'
+
+import { useState } from 'react'
+import { UsernameDialog } from '@/features/rooms/livekit/components/dialogs/UsernameDialog'
+import { SettingsDialog } from '@/features/settings'
+import { OptionsMenuItems } from '@/features/rooms/livekit/components/controls/Options/OptionsMenuItems'
+
+export type DialogState = 'username' | 'settings' | null
 
 export const OptionsButton = () => {
   const { t } = useTranslation('rooms')
+  const [dialogOpen, setDialogOpen] = useState<DialogState>(null)
   return (
-    <MenuTrigger>
-      <Button
-        square
-        legacyStyle
-        aria-label={t('options.buttonLabel')}
-        tooltip={t('options.buttonLabel')}
-      >
-        <RiMore2Line />
-      </Button>
-      <OptionsMenu />
-    </MenuTrigger>
+    <>
+      <Menu>
+        <Button
+          square
+          legacyStyle
+          aria-label={t('options.buttonLabel')}
+          tooltip={t('options.buttonLabel')}
+        >
+          <RiMore2Line />
+        </Button>
+        <OptionsMenuItems onOpenDialog={setDialogOpen} />
+      </Menu>
+      <UsernameDialog
+        isOpen={dialogOpen === 'username'}
+        onOpenChange={(v) => !v && setDialogOpen(null)}
+      />
+      <SettingsDialog
+        isOpen={dialogOpen === 'settings'}
+        onOpenChange={(v) => !v && setDialogOpen(null)}
+      />
+    </>
   )
 }
